+++ tensorflow/lite/toco/model_cmdline_flags.cc
@@ -263,7 +263,7 @@ void ReadModelFlagsFromCommandLineFlags(
     QCHECK(uses_multi_input_flags);
     std::vector<string> mean_values =
         absl::StrSplit(parsed_model_flags.mean_values.value(), ',');
-    QCHECK(mean_values.size() == model_flags->input_arrays_size());
+    QCHECK(static_cast<int>(mean_values.size()) == model_flags->input_arrays_size());
     for (size_t i = 0; i < mean_values.size(); ++i) {
       char* last = nullptr;
       model_flags->mutable_input_arrays(i)->set_mean_value(
@@ -280,7 +280,7 @@ void ReadModelFlagsFromCommandLineFlags(
     QCHECK(uses_multi_input_flags);
     std::vector<string> std_values =
         absl::StrSplit(parsed_model_flags.std_values.value(), ',');
-    QCHECK(std_values.size() == model_flags->input_arrays_size());
+    QCHECK( static_cast<int>(std_values.size()) == model_flags->input_arrays_size());
     for (size_t i = 0; i < std_values.size(); ++i) {
       char* last = nullptr;
       model_flags->mutable_input_arrays(i)->set_std_value(
@@ -298,7 +298,7 @@ void ReadModelFlagsFromCommandLineFlags(
     QCHECK(uses_multi_input_flags);
     std::vector<string> input_data_types =
         absl::StrSplit(parsed_model_flags.input_data_types.value(), ',');
-    QCHECK(input_data_types.size() == model_flags->input_arrays_size());
+    QCHECK(static_cast<int>(input_data_types.size()) == model_flags->input_arrays_size());
     for (size_t i = 0; i < input_data_types.size(); ++i) {
       IODataType type;
       QCHECK(IODataType_Parse(input_data_types[i], &type));
@@ -321,7 +321,7 @@ void ReadModelFlagsFromCommandLineFlags(
     QCHECK(uses_multi_input_flags);
     std::vector<string> input_shapes =
         absl::StrSplit(parsed_model_flags.input_shapes.value(), ':');
-    QCHECK(input_shapes.size() == model_flags->input_arrays_size());
+    QCHECK(static_cast<int>(input_shapes.size()) == model_flags->input_arrays_size());
     for (size_t i = 0; i < input_shapes.size(); ++i) {
       auto* shape = model_flags->mutable_input_arrays(i)->mutable_shape();
       shape->clear_dims();
